Update 0.4.1
1. Fixed a bug cause player cannot join server.
2. Change craft difficulty from integer to float.
3. Add a global multiplier and mod-level multipliers to the config file.
4. Player's experience level now affect the crafting speed (+2% per
level until 200 levels)

+++ src/main/java/sn2/timecraft/ITimeCraftPlayer.java
@@ -8,17 +8,17 @@ public interface ITimeCraftPlayer {
 
 	public boolean isCrafting();
 
-	public void setCraftTime(int craft_time);
+	public void setCraftTime(float craft_time);
 
-	public int getCraftTime();
+	public float getCraftTime();
 
-	public void setCraftPeriod(int craft_period);
+	public void setCraftPeriod(float craft_period);
 
-	public int getCraftPeriod();
+	public float getCraftPeriod();
 	
 	public void stopCraft();
 	
-	public void startCraftWithNewPeriod(int craft_period);
+	public void startCraftWithNewPeriod(float craft_period);
 
 	public boolean tick(ItemStack resultStack);
